==> 00_merge_all_sample_files_with_sample_order.ipynb <==
import pandas as pd 
import pathlib
import os

path_to_20240924_data = "/media/hieunguyen/HNSD01/src/gs-mrd/model_files/10062024/features"
path_to_save_data = "/media/hieunguyen/HNSD01/src/gs-mrd/all_samples/20241209"
os.system(f"mkdir -p {path_to_save_data}")

all_files = [item for item in pathlib.Path(path_to_20240924_data).glob("*features.csv")]
checkdf = pd.read_csv("/media/hieunguyen/HNSD01/src/MRD_GW/release_20240602/check.csv")
predefined_order = checkdf["SampleID"].to_list()
for file in all_files:
    tmpdf = pd.read_csv(file)
    print(tmpdf.shape)
    tmpdf["check"] = tmpdf["SampleID"].apply(lambda x: x.split("-")[1])
    tmpdf.set_index("check", inplace=True)
    tmpdf = tmpdf.loc[predefined_order]
    tmpdf.set_index("SampleID", inplace=True)
    tmpdf = tmpdf.reset_index()
    tmpdf.to_csv(f"{path_to_save_data}/{file.name}", index=False)
